Update method metadata collection location
